+++ packages/twenty-front/src/modules/views/components/EditableFilterDropdownButton.tsx
@@ -1,4 +1,4 @@
-import { useCallback, useEffect } from 'react';
+import { useCallback } from 'react';
 
 import { RecordFilter } from '@/object-record/record-filter/types/RecordFilter';
 import { Dropdown } from '@/ui/layout/dropdown/components/Dropdown';
@@ -6,18 +6,11 @@ import { useDropdown } from '@/ui/layout/dropdown/hooks/useDropdown';
 import { HotkeyScope } from '@/ui/utilities/hotkey/types/HotkeyScope';
 import { EditableFilterChip } from '@/views/components/EditableFilterChip';
 
-import { formatFieldMetadataItemAsFilterDefinition } from '@/object-metadata/utils/formatFieldMetadataItemsAsFilterDefinitions';
 import { ObjectFilterOperandSelectAndInput } from '@/object-record/object-filter-dropdown/components/ObjectFilterOperandSelectAndInput';
-import { fieldMetadataItemIdUsedInDropdownComponentState } from '@/object-record/object-filter-dropdown/states/fieldMetadataItemIdUsedInDropdownComponentState';
-import { filterDefinitionUsedInDropdownComponentState } from '@/object-record/object-filter-dropdown/states/filterDefinitionUsedInDropdownComponentState';
-import { selectedFilterComponentState } from '@/object-record/object-filter-dropdown/states/selectedFilterComponentState';
-import { selectedOperandInDropdownComponentState } from '@/object-record/object-filter-dropdown/states/selectedOperandInDropdownComponentState';
-import { useFilterableFieldMetadataItems } from '@/object-record/record-filter/hooks/useFilterableFieldMetadataItems';
 import { useRemoveRecordFilter } from '@/object-record/record-filter/hooks/useRemoveRecordFilter';
 import { RecordFilterOperand } from '@/object-record/record-filter/types/RecordFilterOperand';
-import { useSetRecoilComponentStateV2 } from '@/ui/utilities/state/component-state/hooks/useSetRecoilComponentStateV2';
+import { EditableFilterDropdownButtonEffect } from '@/views/components/EditableFilterDropdownButtonEffect';
 import { useDeleteCombinedViewFilters } from '@/views/hooks/useDeleteCombinedViewFilters';
-import { isDefined } from 'twenty-shared';
 
 type EditableFilterDropdownButtonProps = {
   viewFilterDropdownId: string;
@@ -30,60 +23,9 @@ export const EditableFilterDropdownButton = ({
   viewFilter,
   hotkeyScope,
 }: EditableFilterDropdownButtonProps) => {
-  const setFilterDefinitionUsedInDropdown = useSetRecoilComponentStateV2(
-    filterDefinitionUsedInDropdownComponentState,
-    viewFilterDropdownId,
-  );
-
-  const setFieldMetadataItemIdUsedInDropdown = useSetRecoilComponentStateV2(
-    fieldMetadataItemIdUsedInDropdownComponentState,
-  );
-
-  const setSelectedOperandInDropdown = useSetRecoilComponentStateV2(
-    selectedOperandInDropdownComponentState,
-    viewFilterDropdownId,
-  );
-
-  const setSelectedFilter = useSetRecoilComponentStateV2(
-    selectedFilterComponentState,
-    viewFilterDropdownId,
-  );
-
-  const { filterableFieldMetadataItems } = useFilterableFieldMetadataItems();
-
   const { closeDropdown } = useDropdown(viewFilterDropdownId);
 
   const { deleteCombinedViewFilter } = useDeleteCombinedViewFilters();
-
-  useEffect(() => {
-    const fieldMetadataItem = filterableFieldMetadataItems.find(
-      (fieldMetadataItem) =>
-        fieldMetadataItem.id === viewFilter.fieldMetadataId,
-    );
-
-    if (!isDefined(fieldMetadataItem)) {
-      return;
-    }
-
-    const filterDefinition = formatFieldMetadataItemAsFilterDefinition({
-      field: fieldMetadataItem,
-    });
-
-    if (isDefined(filterDefinition)) {
-      setFilterDefinitionUsedInDropdown(filterDefinition);
-      setFieldMetadataItemIdUsedInDropdown(filterDefinition.fieldMetadataId);
-      setSelectedOperandInDropdown(viewFilter.operand);
-      setSelectedFilter(viewFilter);
-    }
-  }, [
-    filterableFieldMetadataItems,
-    setFilterDefinitionUsedInDropdown,
-    setFieldMetadataItemIdUsedInDropdown,
-    viewFilter,
-    setSelectedOperandInDropdown,
-    setSelectedFilter,
-    viewFilterDropdownId,
-  ]);
 
   const { removeRecordFilter } = useRemoveRecordFilter();
 
@@ -112,20 +54,26 @@ export const EditableFilterDropdownButton = ({
   }, [viewFilter, deleteCombinedViewFilter, removeRecordFilter]);
 
   return (
-    <Dropdown
-      dropdownId={viewFilterDropdownId}
-      clickableComponent={
-        <EditableFilterChip viewFilter={viewFilter} onRemove={handleRemove} />
-      }
-      dropdownComponents={
-        <ObjectFilterOperandSelectAndInput
-          filterDropdownId={viewFilterDropdownId}
-        />
-      }
-      dropdownHotkeyScope={hotkeyScope}
-      dropdownOffset={{ y: 8, x: 0 }}
-      dropdownPlacement="bottom-start"
-      onClickOutside={handleDropdownClickOutside}
-    />
+    <>
+      <EditableFilterDropdownButtonEffect
+        viewFilterDropdownId={viewFilterDropdownId}
+        viewFilter={viewFilter}
+      />
+      <Dropdown
+        dropdownId={viewFilterDropdownId}
+        clickableComponent={
+          <EditableFilterChip viewFilter={viewFilter} onRemove={handleRemove} />
+        }
+        dropdownComponents={
+          <ObjectFilterOperandSelectAndInput
+            filterDropdownId={viewFilterDropdownId}
+          />
+        }
+        dropdownHotkeyScope={hotkeyScope}
+        dropdownOffset={{ y: 8, x: 0 }}
+        dropdownPlacement="bottom-start"
+        onClickOutside={handleDropdownClickOutside}
+      />
+    </>
   );
 };
