+++ packages/twenty-front/src/modules/views/components/EditableFilterDropdownButtonEffect.tsx
@@ -1,0 +1,74 @@
+import { useEffect } from 'react';
+
+import { formatFieldMetadataItemAsFilterDefinition } from '@/object-metadata/utils/formatFieldMetadataItemsAsFilterDefinitions';
+import { fieldMetadataItemIdUsedInDropdownComponentState } from '@/object-record/object-filter-dropdown/states/fieldMetadataItemIdUsedInDropdownComponentState';
+import { filterDefinitionUsedInDropdownComponentState } from '@/object-record/object-filter-dropdown/states/filterDefinitionUsedInDropdownComponentState';
+import { selectedFilterComponentState } from '@/object-record/object-filter-dropdown/states/selectedFilterComponentState';
+import { selectedOperandInDropdownComponentState } from '@/object-record/object-filter-dropdown/states/selectedOperandInDropdownComponentState';
+import { useFilterableFieldMetadataItems } from '@/object-record/record-filter/hooks/useFilterableFieldMetadataItems';
+import { RecordFilter } from '@/object-record/record-filter/types/RecordFilter';
+import { useSetRecoilComponentStateV2 } from '@/ui/utilities/state/component-state/hooks/useSetRecoilComponentStateV2';
+import { isDefined } from 'twenty-shared';
+
+type EditableFilterDropdownButtonEffectProps = {
+  viewFilterDropdownId: string;
+  viewFilter: RecordFilter;
+};
+
+export const EditableFilterDropdownButtonEffect = ({
+  viewFilterDropdownId,
+  viewFilter,
+}: EditableFilterDropdownButtonEffectProps) => {
+  const setFilterDefinitionUsedInDropdown = useSetRecoilComponentStateV2(
+    filterDefinitionUsedInDropdownComponentState,
+    viewFilterDropdownId,
+  );
+
+  const setFieldMetadataItemIdUsedInDropdown = useSetRecoilComponentStateV2(
+    fieldMetadataItemIdUsedInDropdownComponentState,
+  );
+
+  const setSelectedOperandInDropdown = useSetRecoilComponentStateV2(
+    selectedOperandInDropdownComponentState,
+    viewFilterDropdownId,
+  );
+
+  const setSelectedFilter = useSetRecoilComponentStateV2(
+    selectedFilterComponentState,
+    viewFilterDropdownId,
+  );
+
+  const { filterableFieldMetadataItems } = useFilterableFieldMetadataItems();
+
+  useEffect(() => {
+    const fieldMetadataItem = filterableFieldMetadataItems.find(
+      (fieldMetadataItem) =>
+        fieldMetadataItem.id === viewFilter.fieldMetadataId,
+    );
+
+    if (!isDefined(fieldMetadataItem)) {
+      return;
+    }
+
+    const filterDefinition = formatFieldMetadataItemAsFilterDefinition({
+      field: fieldMetadataItem,
+    });
+
+    if (isDefined(filterDefinition)) {
+      setFilterDefinitionUsedInDropdown(filterDefinition);
+      setFieldMetadataItemIdUsedInDropdown(filterDefinition.fieldMetadataId);
+      setSelectedOperandInDropdown(viewFilter.operand);
+      setSelectedFilter(viewFilter);
+    }
+  }, [
+    filterableFieldMetadataItems,
+    setFilterDefinitionUsedInDropdown,
+    setFieldMetadataItemIdUsedInDropdown,
+    viewFilter,
+    setSelectedOperandInDropdown,
+    setSelectedFilter,
+    viewFilterDropdownId,
+  ]);
+
+  return null;
+};
