+++ packages/twenty-front/src/modules/views/hooks/__tests__/useApplyCurrentViewFiltersToCurrentRecordFilters.test.tsx
@@ -8,6 +8,7 @@ import { useRecoilComponentValueV2 } from '@/ui/utilities/state/component-state/
 import { currentViewIdComponentState } from '@/views/states/currentViewIdComponentState';
 import { ViewFilter } from '@/views/types/ViewFilter';
 import { ViewFilterOperand } from '@/views/types/ViewFilterOperand';
+import { isDefined } from 'packages/twenty-shared/dist';
 import { getJestMetadataAndApolloMocksWrapper } from '~/testing/jest/getJestMetadataAndApolloMocksWrapper';
 import { generatedMockObjectMetadataItems } from '~/testing/mock-data/generatedMockObjectMetadataItems';
 import { useApplyCurrentViewFiltersToCurrentRecordFilters } from '../useApplyCurrentViewFiltersToCurrentRecordFilters';
@@ -15,10 +16,15 @@ import { useApplyCurrentViewFiltersToCurrentRecordFilters } from '../useApplyCur
 jest.mock('@/prefetch/hooks/usePrefetchedData');
 
 describe('useApplyCurrentViewFiltersToCurrentRecordFilters', () => {
-  const mockObjectMetadataItem =
-    generatedMockObjectMetadataItems.find(
-      (item) => item.nameSingular === 'company',
-    ) ?? generatedMockObjectMetadataItems[0];
+  const mockObjectMetadataItem = generatedMockObjectMetadataItems.find(
+    (item) => item.nameSingular === 'company',
+  );
+
+  if (!isDefined(mockObjectMetadataItem)) {
+    throw new Error(
+      'Missing mock object metadata item with name singular "company"',
+    );
+  }
 
   const mockFieldMetadataItem = mockObjectMetadataItem.fields[0];
