+++ packages/twenty-front/src/modules/views/hooks/__tests__/useApplyViewFiltersToCurrentRecordFilters.test.tsx
@@ -6,15 +6,21 @@ import { RecordFilterDefinition } from '@/object-record/record-filter/types/Reco
 import { useRecoilComponentValueV2 } from '@/ui/utilities/state/component-state/hooks/useRecoilComponentValueV2';
 import { ViewFilter } from '@/views/types/ViewFilter';
 import { ViewFilterOperand } from '@/views/types/ViewFilterOperand';
+import { isDefined } from 'packages/twenty-shared/dist';
 import { getJestMetadataAndApolloMocksWrapper } from '~/testing/jest/getJestMetadataAndApolloMocksWrapper';
 import { generatedMockObjectMetadataItems } from '~/testing/mock-data/generatedMockObjectMetadataItems';
 import { useApplyViewFiltersToCurrentRecordFilters } from '../useApplyViewFiltersToCurrentRecordFilters';
 
 describe('useApplyViewFiltersToCurrentRecordFilters', () => {
-  const mockObjectMetadataItem =
-    generatedMockObjectMetadataItems.find(
-      (item) => item.nameSingular === 'company',
-    ) ?? generatedMockObjectMetadataItems[0];
+  const mockObjectMetadataItem = generatedMockObjectMetadataItems.find(
+    (item) => item.nameSingular === 'company',
+  );
+
+  if (!isDefined(mockObjectMetadataItem)) {
+    throw new Error(
+      'Missing mock object metadata item with name singular "company"',
+    );
+  }
 
   const mockFieldMetadataItem = mockObjectMetadataItem.fields[0];
